tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Tools for Raspberry Pi Temperature Logger"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paramiko>=3.5.1",
    "pyyaml>=6.0.2",
    "scp>=0.15.0",
]
